Update README: incompatibilities between this and official builds
This Docker configuration as of 27. March 2017 does not work reliably.
The official build of XenServer relies on a patched CentOS base image
that isn't publically available. This creates incompatibilities between
binaries created with this Docker image and binaries created with
official builds. In can be still used to compile code, though.

Signed-off-by: Christian Lindig <christian.lindig@citrix.com>